Splitted libarea wrapper into separate file
